fixing issue with icons importing
when deploying I got this error
"he target environment doesn't support dynamic import() syntax so it's not possible to use external type 'module' within a script"

Stackoverflow said to improt the icons cdn into the css and that should fix it. The app runs fine locally so we will try that for now.